avaliadores_crew/pyproject.toml
[project]
name = "avaliadores_crew"
version = "0.1.0"
description = "avaliadores-crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.23.0,<1.0.0",
    "celery>=5.3.4,<6.0.0",
    "redis>=5.0.0,<6.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
    "python-multipart>=0.0.6,<0.1.0",
    "pydantic>=2.4.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[project.scripts]
avaliadores_crew = "avaliadores_crew.main:run"
run_crew = "avaliadores_crew.main:run"
train = "avaliadores_crew.main:train"
replay = "avaliadores_crew.main:replay"
test = "avaliadores_crew.main:test"
api = "avaliadores_crew.api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
